requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
backports.zoneinfo==0.2.1
billiard==3.6.4.0
black==23.3.0
cached-property==1.5.2
celery==5.2.7
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==3.2
django-celery-beat==2.4.0
django-timezone-field==5.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.0
flower==2.0.1
gunicorn==22.0.0
humanize==4.6.0
importlib-metadata==4.13.0
kombu==5.2.4
msgpack==1.0.5
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.11.2
platformdirs==4.0.0
prometheus-client==0.17.1
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
redis==5.0.7
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
tornado==6.2
typed-ast==1.5.5
typing_extensions==4.7.1
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
zipp==3.15.0
